Dfsu refactored into smaller pieces
